requirements.txt
faiss-cpu==1.13.2
numpy==2.3.5
pandas==3.0.2
requests==2.33.1
sentence-transformers==5.1.1
torch==2.2.2
tqdm==4.67.3
umap-learn==0.5.12
